src/collect-metrics/requirements.txt
requests
prometheus-client
PyYAML